+++ ocrmypdf/tesseract.py
@@ -6,7 +6,8 @@ import os
 import re
 import shutil
 from functools import lru_cache
-from . import ExitCode, get_program
+from . import ExitCode, get_program, page_number
+from collections import namedtuple
 
 from subprocess import Popen, PIPE, CalledProcessError, \
     TimeoutExpired, check_output, STDOUT
@@ -15,6 +16,10 @@ try:
 except ImportError:
     DEVNULL = open(os.devnull, 'wb')
 
+
+OrientationConfidence = namedtuple(
+    'OrientationConfidence',
+    ('angle', 'confidence'))
 
 HOCR_TEMPLATE = '''<?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
@@ -76,6 +81,53 @@ def languages():
     return set(lang.strip() for lang in langs.splitlines()[1:])
 
 
+def get_orientation(input_file, language: list, timeout: float, log):
+    args_tesseract = [
+        get_program('tesseract'),
+        '-l', '+'.join(language),
+        '-psm', '0',
+        input_file,
+        'stdout'
+    ]
+
+    p = Popen(args_tesseract, close_fds=True, stdout=PIPE, stderr=STDOUT,
+              universal_newlines=True)
+    try:
+        stdout, _ = p.communicate(timeout=timeout)
+    except TimeoutExpired:
+        p.kill()
+        stdout, _ = p.communicate()
+        return OrientationConfidence(angle=0, confidence=0.0)
+    else:
+        osd = {}
+        for line in stdout.splitlines():
+            line = line.strip()
+            parts = line.split(':', maxsplit=2)
+            if len(parts) == 2:
+                osd[parts[0].strip()] = parts[1].strip()
+
+        oc = OrientationConfidence(
+            angle=int(osd['Orientation in degrees']),
+            confidence=float(osd['Orientation confidence']))
+        return oc
+
+
+def tesseract_log_output(log, stdout, input_file):
+    lines = stdout.splitlines()
+    prefix = "{0:4d}: [tesseract] ".format(page_number(input_file))
+    for line in lines:
+        if line.startswith("Tesseract Open Source"):
+            continue
+        elif line.startswith("Warning in pixReadMem"):
+            continue
+        elif 'diacritics' in line:
+            log.warning(prefix + "lots of diacritics - possibly poor OCR")
+        elif line.startswith('OSD: Weak margin'):
+            log.warning(prefix + "unsure about page orientation")
+        else:
+            log.info(prefix + line.strip())
+
+
 def generate_hocr(input_file, output_hocr, language: list, tessconfig: list,
                   timeout: float, pageinfo_getter, pagesegmode: int, log):
 
@@ -94,13 +146,13 @@ def generate_hocr(input_file, output_hocr, language: list, tessconfig: list,
         badxml,
         'hocr'
     ] + tessconfig)
-    p = Popen(args_tesseract, close_fds=True, stdout=PIPE, stderr=PIPE,
+    p = Popen(args_tesseract, close_fds=True, stdout=PIPE, stderr=STDOUT,
               universal_newlines=True)
     try:
-        stdout, stderr = p.communicate(timeout=timeout)
+        stdout, _ = p.communicate(timeout=timeout)
     except TimeoutExpired:
         p.kill()
-        stdout, stderr = p.communicate()
+        stdout, _ = p.communicate()
         # Generate a HOCR file with no recognized text if tesseract times out
         # Temporary workaround to hocrTransform not being able to function if
         # it does not have a valid hOCR file.
@@ -110,11 +162,7 @@ def generate_hocr(input_file, output_hocr, language: list, tessconfig: list,
                 pageinfo['width_pixels'],
                 pageinfo['height_pixels']))
     else:
-        if stdout:
-            log.info(stdout)
-        if stderr:
-            log.error(stderr)
-
+        tesseract_log_output(log, stdout, input_file)
         if p.returncode != 0:
             raise CalledProcessError(p.returncode, args_tesseract)
 
@@ -165,17 +213,14 @@ def generate_pdf(input_image, skip_pdf, output_pdf, language: list,
         os.path.splitext(output_pdf)[0],  # Tesseract appends suffix
         'pdf'
     ] + tessconfig)
-    p = Popen(args_tesseract, close_fds=True, stdout=PIPE, stderr=PIPE,
+    p = Popen(args_tesseract, close_fds=True, stdout=PIPE, stderr=STDOUT,
               universal_newlines=True)
 
     try:
-        stdout, stderr = p.communicate(timeout=timeout)
-        if stdout:
-            log.info(stdout)
-        if stderr:
-            log.error(stderr)
+        stdout, _ = p.communicate()
     except TimeoutExpired:
         p.kill()
         log.info("Tesseract - page timed out")
         shutil.copy(skip_pdf, output_pdf)
-
+    else:
+        tesseract_log_output(log, stdout, input_image)